requirements.txt
pandas
matplotlib
faice==1.5
requests==2.22.0
ruamel.yaml==0.16.7
seaborn==0.10.0
